Refactor and move code
Remove hasDate function as it is needed for the protocol
